backend/requirements.txt
autopep8==1.4.4
config==0.3.9
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==1.11.8
django-bootstrap4==0.0.8
django-cors-headers==3.0.2
django-heroku==0.3.1
django-registration==2.4.1
djangorestframework==3.9.4
gunicorn==19.9.0
Pillow==5.2.0
psycopg2==2.7.5
psycopg2-binary==2.8.3
pycodestyle==2.5.0
python-decouple==3.1
pytz==2019.1
whitenoise==4.1.2
cycler==0.10.0
decorator==4.4.0
imageio==2.5.0
imutils==0.5.2
isort==4.3.21
kiwisolver==1.1.0
lazy-object-proxy==1.4.1
matplotlib==3.1.1
mccabe==0.6.1
networkx==2.3
numpy==1.16.4
opencv-python==4.1.0.25
pylint==2.3.1
pyparsing==2.4.0
python-dateutil==2.8.0
PyWavelets==1.0.3
scikit-image==0.15.0
scipy==1.3.0
six==1.12.0
typed-ast==1.4.0
wrapt==1.11.2